# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MMO Store API - Marketplace for source code and MMO resources
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Email sent (or user not found - same response for security)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token:
                  type: string
                newPassword:
                  type: string
                  minLength: 6
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or expired token

  # Categories
...
  /balance-requests:
    get:
      operationId: listBalanceRequests
      tags: [users]
      summary: List balance requests (admin sees all, user sees own)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of balance requests
          content:
            application/json:
              schema:
                type: object
                required: [requests, total]
                properties:
                  requests:
                    type: array
                    items:
                      $ref: "#/components/schemas/BalanceRequest"
                  total:
                    type: integer
    post:
      operationId: createBalanceRequest
      tags: [users]
      summary: Submit a balance top-up request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BalanceRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceRequest"
  /balance-requests/{id}/approve:
    post:
      operationId: approveBalanceRequest
      tags: [users]
      summary: Approve a balance request (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                adminNote:
                  type: string
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceRequest"
  /balance-requests/{id}/reject:
    post:
      operationId: rejectBalanceRequest
      tags: [users]
      summary: Reject a balance request (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                adminNote:
                  type: string
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceRequest"

  # Stats / Dashboard
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # Auth
    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, username, email, role, balance, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        balance:
          type: number
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string

    # Categories
...
    BalanceRequest:
      type: object
      required: [id, userId, amount, method, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userEmail:
          type: string
        userUsername:
          type: string
        amount:
          type: number
        method:
          type: string
          enum: [bank_transfer, momo, vnpay, zalopay]
        status:
          type: string
          enum: [pending, approved, rejected]
        note:
          type: string
          nullable: true
        adminNote:
          type: string
          nullable: true
        processedBy:
          type: integer
          nullable: true
        processedAt:
          type: string
          nullable: true
        createdAt:
          type: string
    BalanceRequestInput:
      type: object
      required: [amount, method]
      properties:
        amount:
          type: number
          minimum: 50000
        method:
          type: string
          enum: [bank_transfer, momo, vnpay, zalopay]
        note:
          type: string

    # Stats
...