# mmo-store-project/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MMO Store API - Marketplace for source code and MMO resources
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  username: string;
  email: string;
  role: UserRole;
  balance: number;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  iconName?: string | null;
  productCount?: number;
}

export interface CategoryInput {
  name: string;
  description?: string;
  iconName?: string;
}

export type ProductStatus = typeof ProductStatus[keyof typeof ProductStatus];


export const ProductStatus = {
  active: 'active',
  inactive: 'inactive',
  draft: 'draft',
} as const;

export interface Product {
  id: number;
  name: string;
  slug: string;
  description?: string;
  /** @nullable */
  shortDescription?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  status: ProductStatus;
  featured?: boolean;
  viewCount: number;
  soldCount: number;
  rating: number;
  reviewCount?: number;
  tags?: string[];
  techStack?: string[];
  createdAt: string;
}

export interface Review {
  id: number;
  productId: number;
  userId: number;
  /** @nullable */
  username?: string | null;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @nullable */
  comment?: string | null;
  createdAt: string;
}

export interface ProductDetail {
  id: number;
  name: string;
  slug: string;
  description?: string;
  /** @nullable */
  shortDescription?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  imageUrls?: string[];
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  status: string;
  featured?: boolean;
  viewCount: number;
  soldCount: number;
  rating: number;
  reviewCount?: number;
  tags?: string[];
  techStack?: string[];
  features?: string[];
  /** @nullable */
  demoUrl?: string | null;
  /** @nullable */
  downloadUrl?: string | null;
  createdAt: string;
  reviews: Review[];
}

export interface ProductList {
  products: Product[];
  total: number;
  page: number;
  limit: number;
  totalPages?: number;
}

export type ProductInputStatus = typeof ProductInputStatus[keyof typeof ProductInputStatus];


export const ProductInputStatus = {
  active: 'active',
  inactive: 'inactive',
  draft: 'draft',
} as const;

export interface ProductInput {
  name: string;
  description?: string;
  shortDescription?: string;
  price: number;
  originalPrice?: number;
  thumbnailUrl?: string;
  imageUrls?: string[];
  categoryId: number;
  featured?: boolean;
  status?: ProductInputStatus;
  tags?: string[];
  techStack?: string[];
  features?: string[];
  demoUrl?: string;
  downloadUrl?: string;
}

export type ProductUpdateStatus = typeof ProductUpdateStatus[keyof typeof ProductUpdateStatus];


export const ProductUpdateStatus = {
  active: 'active',
  inactive: 'inactive',
  draft: 'draft',
} as const;

export interface ProductUpdate {
  name?: string;
  description?: string;
  shortDescription?: string;
  price?: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  thumbnailUrl?: string | null;
  imageUrls?: string[];
  categoryId?: number;
  featured?: boolean;
  status?: ProductUpdateStatus;
  tags?: string[];
  techStack?: string[];
  features?: string[];
  /** @nullable */
  demoUrl?: string | null;
  /** @nullable */
  downloadUrl?: string | null;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  completed: 'completed',
  cancelled: 'cancelled',
  refunded: 'refunded',
} as const;

export interface Order {
  id: number;
  userId: number;
  productId: number;
  /** @nullable */
  productName?: string | null;
  /** @nullable */
  productThumbnail?: string | null;
  amount: number;
  status: OrderStatus;
  /** @nullable */
  downloadUrl?: string | null;
  createdAt: string;
}

export interface OrderInput {
  productId: number;
}

export interface ReviewInput {
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  comment?: string;
}

export interface Post {
  id: number;
  title: string;
  slug: string;
  /** @nullable */
  excerpt?: string | null;
  content: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  authorName?: string | null;
  viewCount?: number;
  tags?: string[];
  createdAt: string;
}

export interface PostList {
  posts: Post[];
  total: number;
  page: number;
  limit: number;
  totalPages?: number;
}

export interface PostInput {
  title: string;
  excerpt?: string;
  content: string;
  thumbnailUrl?: string;
  tags?: string[];
}

export type StoreStatsCategoryCountsItem = {
  categoryName?: string;
  count?: number;
};

export interface StoreStats {
  totalProducts: number;
  totalOrders: number;
  totalUsers: number;
  totalRevenue: number;
  topProducts: Product[];
  recentOrders: Order[];
  categoryCounts?: StoreStatsCategoryCountsItem[];
}

export interface Download {
  orderId: number;
  productId: number;
  productName: string;
  /** @nullable */
  productThumbnail?: string | null;
  downloadUrl: string;
  purchasedAt: string;
}

export interface UserStats {
  purchaseCount: number;
  totalSpent: number;
  downloads: Download[];
}

export type BalanceInputMethod = typeof BalanceInputMethod[keyof typeof BalanceInputMethod];


export const BalanceInputMethod = {
  bank_transfer: 'bank_transfer',
  momo: 'momo',
  vnpay: 'vnpay',
  zalopay: 'zalopay',
} as const;

export interface BalanceInput {
  amount: number;
  method?: BalanceInputMethod;
}

export interface ProfileUpdate {
  username?: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export type ListProductsParams = {
categoryId?: number;
search?: string;
minPrice?: number;
maxPrice?: number;
sortBy?: ListProductsSortBy;
page?: number;
limit?: number;
featured?: boolean;
};

export type ListProductsSortBy = typeof ListProductsSortBy[keyof typeof ListProductsSortBy];


export const ListProductsSortBy = {
  newest: 'newest',
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  popular: 'popular',
  rating: 'rating',
} as const;

export type ListPostsParams = {
page?: number;
limit?: number;
};